backend/requirements.txt
# FinBot Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LLM & LangChain
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Document Parsing
docling>=2.0.0
docling-core>=2.0.0

# Vector Database
qdrant-client>=1.12.0

# Embeddings
sentence-transformers>=3.0.0

# Semantic Routing
semantic-router>=0.1.0

# Authentication
pyjwt>=2.9.0
bcrypt>=4.2.0

# Database
aiosqlite>=0.20.0

# Evaluation
ragas>=0.2.0
datasets>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
